healthchatbot-backend/requirements.txt
# 🚦 Person A - Dependencies
fastapi==0.110.0
fastapi==0.110.0
uvicorn[standard]==0.29.0
pymongo==4.3.3
motor==3.1.1

python-decouple==3.8
pydantic==2.5.0
twilio==8.10.0
langdetect==1.0.9
geopy==2.4.1
schedule==1.2.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2

python-decouple
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
twilio==8.10.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
langdetect==1.0.9
asyncio==3.4.3
pathlib2==2.3.7
